- Updated to use Job table which has several Roles - Created list Employees, EmployeeDetails pages

+++ SavvyControlDesk/src/main/java/com/savvy/v1/controldesk/converters/EmployeeConverter.java
@@ -17,7 +17,7 @@ public class EmployeeConverter implements Converter {
 	
 	 @Override
 	    public Object getAsObject(FacesContext context, UIComponent component, String value) {
-	        Employee e = employeeDao.getById(Long.parseLong(value));
+	        Employee e = employeeDao.getById((int) Long.parseLong(value));
 	        return e;
 	    }
